backend/requirements.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
six==1.15.0
virtualenv==20.4.2
Flask
flask_cors
psycopg2-binary
